feat: implement BettingPanel and CoinTossGameControls components, refactor CoinTossGame to utilize new controls and update GameFrame structure
